Added a reels behaviour
