+++ src/Screens/DetailsScreen.tsx
@@ -1,6 +1,5 @@
 import React, {useCallback, useState} from 'react';
 import {
-  Dimensions,
   FlatList,
   Image,
   Pressable,
@@ -10,8 +9,8 @@ import {
   useWindowDimensions,
 } from 'react-native';
 import DetailsHeader from '../components/DetailsHeader';
-import {gridImages} from '../assets/mock-data';
-import type {gridImagesType} from '../assets/mock-data';
+import {gridImages} from '../assets/grid-mock-data';
+import type {gridImagesType} from '../assets/grid-mock-data';
 import {formatNumber} from '../util/NumberFormatter';
 
 const DetailsScreen = ({navigation}: any) => {
