+++ src/Screens/HomeScreen.tsx
@@ -1,34 +1,16 @@
 import React from 'react';
-import {ImageBackground, StyleSheet, View} from 'react-native';
+import {StyleSheet, View} from 'react-native';
 import LinearGradient from 'react-native-linear-gradient';
-import HomeHeader from '../components/HomeHeader';
-import {Hero} from '../components/Hero';
+import Feed from '../components/Feed';
 
 function HomeScreen() {
   return (
     <>
       <LinearGradient
         colors={['rgba(0, 0, 0, 0.8)', 'transparent']}
-        style={{
-          position: 'absolute',
-          top: 0,
-          left: 0,
-          zIndex: 555,
-          height: 64,
-          width: '100%',
-        }}
+        style={styles.gradient}
       />
-      <View style={styles.container}>
-        <ImageBackground
-          source={require('../assets/breakingbad.jpeg')}
-          resizeMode="cover"
-          style={styles.imageBackground}>
-          <View style={styles.innerContainer}>
-            <HomeHeader />
-            <Hero />
-          </View>
-        </ImageBackground>
-      </View>
+      <Feed />
     </>
   );
 }
@@ -36,38 +18,12 @@ function HomeScreen() {
 export default HomeScreen;
 
 const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-  },
-  innerContainer: {
-    flex: 1,
-    justifyContent: 'space-between',
-    alignItems: 'center',
-  },
-  imageBackground: {
-    zIndex: 1,
-    flex: 1,
-    height: '100%',
+  gradient: {
+    position: 'absolute',
+    top: 0,
+    left: 0,
+    zIndex: 4,
+    height: 64,
     width: '100%',
-  },
-  image: {
-    width: 100,
-    height: 60,
-  },
-  heroContainer: {
-    zIndex: 5,
-    width: '90%',
-    paddingBottom: 60,
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'flex-end',
-  },
-  icons: {
-    gap: 15,
-  },
-  text: {
-    textAlign: 'center',
-    fontSize: 12,
-    color: 'white',
   },
 });
