+++ src/assets/posts-mock-data.ts
@@ -1,0 +1,56 @@
+export const posts = [
+  {
+    id: 2,
+    backgroundSource: require('../assets/breakingbad.jpeg'),
+    title: 'Breaking Bad',
+    logoSource: require('../assets/breaking-bad-logo.png'),
+  },
+  {
+    id: 3,
+    backgroundSource: require('../assets/breakingbad2.jpeg'),
+    title: 'Breaking Bad',
+    logoSource: require('../assets/breaking-bad-logo.png'),
+  },
+  {
+    id: 4,
+    backgroundSource: require('../assets/breakingbad.jpeg'),
+    title: 'Breaking Bad',
+    logoSource: require('../assets/breaking-bad-logo.png'),
+  },
+  {
+    id: 5,
+    backgroundSource: require('../assets/breakingbad2.jpeg'),
+    title: 'Breaking Bad',
+    logoSource: require('../assets/breaking-bad-logo.png'),
+  },
+  {
+    id: 6,
+    backgroundSource: require('../assets/breakingbad.jpeg'),
+    title: 'Breaking Bad',
+    logoSource: require('../assets/breaking-bad-logo.png'),
+  },
+  {
+    id: 7,
+    backgroundSource: require('../assets/breakingbad2.jpeg'),
+    title: 'Breaking Bad',
+    logoSource: require('../assets/breaking-bad-logo.png'),
+  },
+  {
+    id: 8,
+    backgroundSource: require('../assets/breakingbad.jpeg'),
+    title: 'Breaking Bad',
+    logoSource: require('../assets/breaking-bad-logo.png'),
+  },
+  {
+    id: 9,
+    backgroundSource: require('../assets/breakingbad2.jpeg'),
+    title: 'Breaking Bad',
+    logoSource: require('../assets/breaking-bad-logo.png'),
+  },
+  {
+    id: 10,
+    backgroundSource: require('../assets/breakingbad.jpeg'),
+    title: 'Breaking Bad',
+    logoSource: require('../assets/breaking-bad-logo.png'),
+  },
+];
